users_service/requirements.txt
aiokafka==0.12.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.0
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datadog==0.50.1
Django==5.1.1
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
kafka-python==2.0.2
mongoengine==0.29.1
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
